requirements.txt
# ===============================================================
# 📦 Core Dependencies
# ===============================================================
Django==5.1
djangorestframework==3.15.2
django-environ==0.11.2
django-redis==5.4.0
psycopg2-binary==2.9.9
drf-spectacular==0.27.1
djangorestframework-simplejwt==5.3.1
drf-nested-routers

# ===============================================================
# ⚙️ Background Jobs / Scheduling
# ===============================================================
APScheduler==3.10.4

# ===============================================================
# 🧰 Logging / Utilities
# ===============================================================
structlog==24.1.0

# ===============================================================
# 🧪 Testing / Dev Tools
# ===============================================================
factory_boy==3.3.0
Faker==25.0.0
pytest
pytest-django
pytest-cov
black
isort
pre-commit
